test: move all cypress comment in cypress folder
