+++ src/components/Cypress/color-assertions.vue
