+++ src/components/__tests__/Cypress/QuasarButton.cy.js
@@ -1,4 +1,4 @@
-import QuasarButton from '../QuasarButton.vue';
+import QuasarButton from '../../Cypress/QuasarButton.vue';
 
 describe('QuasarButton', () => {
   it('renders a message', () => {
